Move include and define files to configprovider
Since they're only used by the other files in that folder, they should
also be there
