+++ Extension/src/extension/configprovider/dynamicconfiggenerator.ts
@@ -7,8 +7,8 @@ import { Workbench } from "../../iar/tools/workbench";
 import { Config } from "../../iar/project/config";
 import { ConfigurationCache, SimpleConfigurationCache } from "./configurationcache";
 import * as Vscode from "vscode";
-import { IncludePath } from "../../iar/project/includepath";
-import { Define } from "../../iar/project/define";
+import { IncludePath } from "./includepath";
+import { Define } from "./define";
 import { join } from "path";
 import { spawn } from "child_process";
 import * as readline from "readline";
